PhatLoots plugin
Thanks for this awsome plugin! CityWorld is ace.

I added support for PhatLoots plugin when CityWorld creats chests in a sewer vault.

The master branch is broken when creating a world from scratch (before applying these changes).

These changes compile against the master branch and work as intended.

If anybody is interested I have a set of patches backported to the CityWorld v0.61 release.